backend/requirements.txt
fastapi[all]
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-dotenv
psycopg2-binary
